add vlfuse file
